+++ internetarchive/api.py
@@ -487,7 +487,7 @@ def search_items(query,
                                         request_kwargs=request_kwargs)
 
 
-def configure(username=None, password=None):
+def configure(username=None, password=None, config_file=None):
     """Configure internetarchive with your Archive.org credentials.
 
     :type username: str
@@ -502,7 +502,7 @@ def configure(username=None, password=None):
     """
     username = input('Email address: ') if not username else username
     password = getpass('Password: ') if not password else password
-    config_file_path = config_module.write_config_file(username, password)
+    config_file_path = config_module.write_config_file(username, password, config_file)
     print('\nConfig saved to: {0}'.format(config_file_path))